moved notifications service into service folder
